Moved create event to its own component folder
